*appengine -> 1.9.10, *django -> 1.5.8; *songs, my apps +using cache

+++ django/test/simple.py
@@ -1,54 +1,51 @@
-import sys
-import signal
+import unittest as real_unittest
 
 from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
 from django.db.models import get_app, get_apps
 from django.test import _doctest as doctest
 from django.test.utils import setup_test_environment, teardown_test_environment
-from django.test.testcases import OutputChecker, DocTestRunner, TestCase
+from django.test.testcases import OutputChecker, DocTestRunner
 from django.utils import unittest
+from django.utils.importlib import import_module
+from django.utils.module_loading import module_has_submodule
+
+__all__ = ('DjangoTestSuiteRunner')
 
 # The module name for tests outside models.py
 TEST_MODULE = 'tests'
 
 doctestOutputChecker = OutputChecker()
 
-class DjangoTestRunner(unittest.TextTestRunner):
-    def __init__(self, *args, **kwargs):
-        import warnings
-        warnings.warn(
-            "DjangoTestRunner is deprecated; it's functionality is indistinguishable from TextTestRunner",
-            PendingDeprecationWarning
-        )
-        super(DjangoTestRunner, self).__init__(*args, **kwargs)
 
 def get_tests(app_module):
+    parts = app_module.__name__.split('.')
+    prefix, last = parts[:-1], parts[-1]
     try:
-        app_path = app_module.__name__.split('.')[:-1]
-        test_module = __import__('.'.join(app_path + [TEST_MODULE]), {}, {}, TEST_MODULE)
-    except ImportError, e:
+        test_module = import_module('.'.join(prefix + [TEST_MODULE]))
+    except ImportError:
         # Couldn't import tests.py. Was it due to a missing file, or
         # due to an import error in a tests.py that actually exists?
-        import os.path
-        from imp import find_module
-        try:
-            mod = find_module(TEST_MODULE, [os.path.dirname(app_module.__file__)])
-        except ImportError:
-            # 'tests' module doesn't exist. Move on.
+        # app_module either points to a models.py file, or models/__init__.py
+        # Tests are therefore either in same directory, or one level up
+        if last == 'models':
+            app_root = import_module('.'.join(prefix))
+        else:
+            app_root = app_module
+
+        if not module_has_submodule(app_root, TEST_MODULE):
             test_module = None
         else:
-            # The module exists, so there must be an import error in the
-            # test module itself. We don't need the module; so if the
-            # module was a single file module (i.e., tests.py), close the file
-            # handle returned by find_module. Otherwise, the test module
-            # is a directory, and there is nothing to close.
-            if mod[0]:
-                mod[0].close()
+            # The module exists, so there must be an import error in the test
+            # module itself.
             raise
     return test_module
 
+
 def build_suite(app_module):
-    "Create a complete Django test suite for the provided application module"
+    """
+    Create a complete Django test suite for the provided application module.
+    """
     suite = unittest.TestSuite()
 
     # Load unit and doctests in the models.py module. If module has
@@ -56,7 +53,8 @@ def build_suite(app_module):
     if hasattr(app_module, 'suite'):
         suite.addTest(app_module.suite())
     else:
-        suite.addTest(unittest.defaultTestLoader.loadTestsFromModule(app_module))
+        suite.addTest(unittest.defaultTestLoader.loadTestsFromModule(
+            app_module))
         try:
             suite.addTest(doctest.DocTestSuite(app_module,
                                                checker=doctestOutputChecker,
@@ -74,25 +72,29 @@ def build_suite(app_module):
         if hasattr(test_module, 'suite'):
             suite.addTest(test_module.suite())
         else:
-            suite.addTest(unittest.defaultTestLoader.loadTestsFromModule(test_module))
+            suite.addTest(unittest.defaultTestLoader.loadTestsFromModule(
+                test_module))
             try:
-                suite.addTest(doctest.DocTestSuite(test_module,
-                                                   checker=doctestOutputChecker,
-                                                   runner=DocTestRunner))
+                suite.addTest(doctest.DocTestSuite(
+                    test_module, checker=doctestOutputChecker,
+                    runner=DocTestRunner))
             except ValueError:
                 # No doc tests in tests.py
                 pass
     return suite
 
+
 def build_test(label):
-    """Construct a test case with the specified label. Label should be of the
+    """
+    Construct a test case with the specified label. Label should be of the
     form model.TestClass or model.TestClass.test_method. Returns an
     instantiated test or test suite corresponding to the label provided.
 
     """
     parts = label.split('.')
     if len(parts) < 2 or len(parts) > 3:
-        raise ValueError("Test label '%s' should be of the form app.TestCase or app.TestCase.test_method" % label)
+        raise ValueError("Test label '%s' should be of the form app.TestCase "
+                         "or app.TestCase.test_method" % label)
 
     #
     # First, look for TestCase instances with a name that matches
@@ -107,12 +109,15 @@ def build_test(label):
             TestClass = getattr(test_module, parts[1], None)
 
     try:
-        if issubclass(TestClass, unittest.TestCase):
+        if issubclass(TestClass, (unittest.TestCase, real_unittest.TestCase)):
             if len(parts) == 2: # label is app.TestClass
                 try:
-                    return unittest.TestLoader().loadTestsFromTestCase(TestClass)
+                    return unittest.TestLoader().loadTestsFromTestCase(
+                        TestClass)
                 except TypeError:
-                    raise ValueError("Test label '%s' does not refer to a test class" % label)
+                    raise ValueError(
+                        "Test label '%s' does not refer to a test class"
+                        % label)
             else: # label is app.TestClass.test_method
                 return TestClass(parts[2])
     except TypeError:
@@ -133,7 +138,8 @@ def build_test(label):
             for test in doctests:
                 if test._dt_test.name in (
                         '%s.%s' % (module.__name__, '.'.join(parts[1:])),
-                        '%s.__test__.%s' % (module.__name__, '.'.join(parts[1:]))):
+                        '%s.__test__.%s' % (
+                            module.__name__, '.'.join(parts[1:]))):
                     tests.append(test)
         except ValueError:
             # No doctests found.
@@ -145,6 +151,7 @@ def build_test(label):
 
     # Construct a suite out of the tests that matched.
     return unittest.TestSuite(tests)
+
 
 def partition_suite(suite, classes, bins):
     """
@@ -167,14 +174,15 @@ def partition_suite(suite, classes, bins):
             else:
                 bins[-1].addTest(test)
 
+
 def reorder_suite(suite, classes):
     """
     Reorders a test suite by test type.
 
-    classes is a sequence of types
+    `classes` is a sequence of types
 
-    All tests of type clases[0] are placed first, then tests of type classes[1], etc.
-    Tests with no match in classes are placed last.
+    All tests of type classes[0] are placed first, then tests of type
+    classes[1], etc. Tests with no match in classes are placed last.
     """
     class_count = len(classes)
     bins = [unittest.TestSuite() for i in range(class_count+1)]
@@ -182,6 +190,48 @@ def reorder_suite(suite, classes):
     for i in range(class_count):
         bins[0].addTests(bins[i+1])
     return bins[0]
+
+
+def dependency_ordered(test_databases, dependencies):
+    """
+    Reorder test_databases into an order that honors the dependencies
+    described in TEST_DEPENDENCIES.
+    """
+    ordered_test_databases = []
+    resolved_databases = set()
+
+    # Maps db signature to dependencies of all it's aliases
+    dependencies_map = {}
+
+    # sanity check - no DB can depend on it's own alias
+    for sig, (_, aliases) in test_databases:
+        all_deps = set()
+        for alias in aliases:
+            all_deps.update(dependencies.get(alias, []))
+        if not all_deps.isdisjoint(aliases):
+            raise ImproperlyConfigured(
+                "Circular dependency: databases %r depend on each other, "
+                "but are aliases." % aliases)
+        dependencies_map[sig] = all_deps
+
+    while test_databases:
+        changed = False
+        deferred = []
+
+        # Try to find a DB that has all it's dependencies met
+        for signature, (db_name, aliases) in test_databases:
+            if dependencies_map[signature].issubset(resolved_databases):
+                resolved_databases.update(aliases)
+                ordered_test_databases.append((signature, (db_name, aliases)))
+                changed = True
+            else:
+                deferred.append((signature, (db_name, aliases)))
+
+        if not changed:
+            raise ImproperlyConfigured(
+                "Circular dependency in TEST_DEPENDENCIES")
+        test_databases = deferred
+    return ordered_test_databases
 
 
 class DjangoTestSuiteRunner(object):
@@ -213,37 +263,81 @@ class DjangoTestSuiteRunner(object):
             for test in extra_tests:
                 suite.addTest(test)
 
-        return reorder_suite(suite, (TestCase,))
+        return reorder_suite(suite, (unittest.TestCase,))
 
     def setup_databases(self, **kwargs):
-        from django.db import connections
-        old_names = []
-        mirrors = []
+        from django.db import connections, DEFAULT_DB_ALIAS
+
+        # First pass -- work out which databases actually need to be created,
+        # and which ones are test mirrors or duplicate entries in DATABASES
+        mirrored_aliases = {}
+        test_databases = {}
+        dependencies = {}
+        default_sig = connections[DEFAULT_DB_ALIAS].creation.test_db_signature()
         for alias in connections:
             connection = connections[alias]
-            # If the database is a test mirror, redirect it's connection
-            # instead of creating a test database.
             if connection.settings_dict['TEST_MIRROR']:
-                mirrors.append((alias, connection))
-                mirror_alias = connection.settings_dict['TEST_MIRROR']
-                connections._connections[alias] = connections[mirror_alias]
+                # If the database is marked as a test mirror, save
+                # the alias.
+                mirrored_aliases[alias] = (
+                    connection.settings_dict['TEST_MIRROR'])
             else:
-                old_names.append((connection, connection.settings_dict['NAME']))
-                connection.creation.create_test_db(self.verbosity, autoclobber=not self.interactive)
+                # Store a tuple with DB parameters that uniquely identify it.
+                # If we have two aliases with the same values for that tuple,
+                # we only need to create the test database once.
+                item = test_databases.setdefault(
+                    connection.creation.test_db_signature(),
+                    (connection.settings_dict['NAME'], set())
+                )
+                item[1].add(alias)
+
+                if 'TEST_DEPENDENCIES' in connection.settings_dict:
+                    dependencies[alias] = (
+                        connection.settings_dict['TEST_DEPENDENCIES'])
+                else:
+                    if alias != DEFAULT_DB_ALIAS and connection.creation.test_db_signature() != default_sig:
+                        dependencies[alias] = connection.settings_dict.get(
+                            'TEST_DEPENDENCIES', [DEFAULT_DB_ALIAS])
+
+        # Second pass -- actually create the databases.
+        old_names = []
+        mirrors = []
+
+        for signature, (db_name, aliases) in dependency_ordered(
+            test_databases.items(), dependencies):
+            test_db_name = None
+            # Actually create the database for the first connection
+
+            for alias in aliases:
+                connection = connections[alias]
+                if test_db_name is None:
+                    test_db_name = connection.creation.create_test_db(
+                            self.verbosity, autoclobber=not self.interactive)
+                    destroy = True
+                else:
+                    connection.settings_dict['NAME'] = test_db_name
+                    destroy = False
+                old_names.append((connection, db_name, destroy))
+
+        for alias, mirror_alias in mirrored_aliases.items():
+            mirrors.append((alias, connections[alias].settings_dict['NAME']))
+            connections[alias].settings_dict['NAME'] = (
+                connections[mirror_alias].settings_dict['NAME'])
+
         return old_names, mirrors
 
     def run_suite(self, suite, **kwargs):
-        return unittest.TextTestRunner(verbosity=self.verbosity, failfast=self.failfast).run(suite)
+        return unittest.TextTestRunner(
+            verbosity=self.verbosity, failfast=self.failfast).run(suite)
 
     def teardown_databases(self, old_config, **kwargs):
-        from django.db import connections
+        """
+        Destroys all the non-mirror databases.
+        """
         old_names, mirrors = old_config
-        # Point all the mirrors back to the originals
-        for alias, connection in mirrors:
-            connections._connections[alias] = connection
-        # Destroy all the non-mirror databases
-        for connection, old_name in old_names:
-            connection.creation.destroy_test_db(old_name, self.verbosity)
+        for connection, old_name, destroy in old_names:
+            if destroy:
+                connection.creation.destroy_test_db(old_name, self.verbosity)
 
     def teardown_test_environment(self, **kwargs):
         unittest.removeHandler()
@@ -278,12 +372,3 @@ class DjangoTestSuiteRunner(object):
         self.teardown_databases(old_config)
         self.teardown_test_environment()
         return self.suite_result(suite, result)
-
-def run_tests(test_labels, verbosity=1, interactive=True, failfast=False, extra_tests=None):
-    import warnings
-    warnings.warn(
-        'The run_tests() test runner has been deprecated in favor of DjangoTestSuiteRunner.',
-        DeprecationWarning
-    )
-    test_runner = DjangoTestSuiteRunner(verbosity=verbosity, interactive=interactive, failfast=failfast)
-    return test_runner.run_tests(test_labels, extra_tests=extra_tests)
